requirements.txt
glfw
PyOpenGL